Add invoice list into the resolvers
Extracted the generic component into its own file. It seems to have a problem with GraphQL aliases
on apollo-link-state for now, so leaving the original names of the JSON instead as props.
It should be updated later.
There is also a fragment but not connected yet since it doesn't work on the main library.
It should be used in the future to abstract better between what needs a component and what the
container is aware of.