Fix storybook and stories
With the new router, warnings are triggered when <Route> or <Link> is used
without having <Router> above in the tree. With storybook-react-router
it fixes this problem and could enable also to link our stories between
them with storybook-link in the future.
This commit fixes also the snapshots that are obsoletes.