Add first tests for subscription component
This is the first steps towards a fully tested app!
Add test for the regular monthly subscription plan
Add test that tries a wrong coupon before subscribing with the right one.

These tests should be changed later, as it'll be better to simplify some code
that seems overkill or redundant.