Add a signin integration test
This test asserts the component works well when the user tries to log in with:
- no data
- wrong password
- correct credentials
JSDOM doesn't support LocalStorage, added a mockup.
The mock prevents the data to be refetched so it needs to be specified in the main component.
This will also pave the future removal of the HoodieApi dependency.